RSS Feed Subscription Integration (#4)
# RSS Feed Integration
This Feature introduces the ability for a RAG workspace to subscribe to an RSS Feed, check for new posts daily and sends new posts for Website Crawling for the RAG Workspace. 

The functionality leverages existing design, treating an RSS Subscription as a type of DocumentItem in the UI. Clicking on an RSS Subscription takes the user to a new RSS Feed page which shows details of the RSS Subscription including Feed URL, last check (it automates daily polling), subscription status (enabled/disabled). 
The RSS Feed screen includes a table that lists all the posts detected from the RSS feed, when it was added to the workspace and if it's pending or sent for processing by the website crawler. Users can toggle a subscription on or off.

Since the functionality leverages the existing Crawl Websites functionality, when a post is crawled, it will show the website in the Crawl Websites list. 

If a workspace is deleted, all schedules are removed and all associated data in the RSS Feed DynamoDB table will be removed.